+++ docs/releases.md
@@ -13,8 +13,9 @@ This is a documentation of upcoming releases. Only minor and major releases are 
 
 ### Progress
 
-- [ ] Add lifecycle hooks.
-- [ ] Add customized component as a parameter to [`createComponent`](api/create-component/create-component.md).
+- [x] Add lifecycle hooks.
+- [x] Add customized component as a parameter to [`createComponent`](api/create-component/create-component.md).
+- [x] Released
 
 ### Release Date